RopeAndReelingRemoved
Rest in peace harpoon rope. I hope you are in a better place now. In all seriousness, deleted the rope and rope spring functionality. Removed the functionality for the harpoon to retract. Instead switched to a reload. This includes changes to the player manager to accordingly. Integrated object pooling into the harpoon projectile.