UniversalManagerReformatting
Subscribes all universal managers to a virtual function for subscribing to gameplay events. Previously it was annoying for universal managers to subscribe to gameplay managers events as it couldn't do so through start/awake/onenable because that wouldn't work if you started on a non gameplay scene. Also added some comments and renamed some events to fit naming conventions