Ambience Adjustments
Adjusted the AmbienceManager and FmodAmbienceEvents to be more in line with the other managers. I don't have start in any other managers, reason being is that the core manager for both the universal manager and gameplay manager call SetUpMainManager on all their associated managers. Also adjusted the FmodAmbienceManager to set up the instance similarly to the managers