vs-ux-menu-overhaul
Overhauled the title screen's menus to be consistent with the pause menu screens, and adjusted the pause menu elements' spacing to be consistent with the settings screen.

Jira Issue: https://bradleycapstone.atlassian.net/browse/LV-619

Changes can be observed on the title screen. Secondary changes can be observed in the pause screen.