Fixed Issue of enemy following player out of room
Reason why is that the enemy was childed to the room, and so when the room was checking for if the player enetered the room it was taking into account both the room collider and the enemy collider. So it would increase the collision counter by 2 rather than 1. Had to make some minor adjustments to the enemy behavior to make that work. The enemy is now a child of the attack rather than the room