ImplementedSomeHorrorMoments
Implemented the Light Change and Slytherin horror moments. Also sort of implemented Thump In The Night but the sound is just the monster screech as the sounds are not in. Also tried starting on the chase sequence, but currently no way to get to the final maze part. Also adjusted the boat sway to be less extreme.