Updated SeedPod Textures
Okay, so, I have made the seed pod textures much brighter than it was before, however in areas where the lighting is dim, it is still reading black. I checked all the other prefabs that were placed in the scene as well, they have all been updated with the new textures. Only caveat is that you can see the ones in better lit areas rather than the others.