Trigger Adjustments
Adjusted the Trigger at the start of the Chase so that the dialogue lines up with the cutscene.

Swapped the placement of the Boat Bell Toll and Line about the boat as it felt it would make more sense for the boat to make the sound before the player comments.